added an exceptions module . created multiple exception files in it to make it easier to maintain
