+++ sark/exceptions/base_exceptions.py
@@ -1,0 +1,2 @@
+class SarkError(Exception):
+    pass
