+++ sark/exceptions/gui_exceptions.py
@@ -1,0 +1,20 @@
+from base_exceptions import *
+
+
+class SarkGuiError(SarkError):
+    pass
+
+
+class SarkMenuError(SarkGuiError):
+    pass
+
+class MenuAlreadyExists(SarkMenuError):
+    pass
+
+
+class MenuNotFound(SarkMenuError):
+    pass
+
+
+class FormNotFound(SarkGuiError):
+    pass
