+++ sark/exceptions/phrase_exceptions.py
@@ -1,0 +1,20 @@
+from base_exceptions import *
+
+
+class SarkOperandError(SarkError):
+    pass
+
+class SarkOperandWithoutReg(SarkOperandError):
+    pass
+
+class PhraseError(SarkOperandError):
+    pass
+
+class OperandNotPhrase(PhraseError):
+    pass
+
+class PhraseNotSupported(PhraseError):
+    pass
+
+class PhraseProcessorNotSupported(PhraseNotSupported):
+    pass
