+++ sark/exceptions/segment_exeptions.py
@@ -1,0 +1,13 @@
+from base_exceptions import *
+
+class SegmentError(SarkError):
+    pass
+
+
+class NoMoreSegments(SegmentError):
+    pass
+
+
+class InvalidBitness(SegmentError):
+    pass
+
