+++ sark/exceptions/switch_exceptions.py
@@ -1,0 +1,9 @@
+from base_exceptions import *
+
+class SarkSwitchError(SarkError):
+    pass
+
+
+class SarkNotASwitch(SarkSwitchError):
+    pass
+
